airflowDAGs/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
alembic==1.14.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.5.2
apache-airflow==2.10.5
apache-airflow-providers-celery==3.8.2
apache-airflow-providers-common-compat==1.2.0
apache-airflow-providers-common-io==1.4.1
apache-airflow-providers-common-sql==1.17.1
apache-airflow-providers-fab==1.4.0
apache-airflow-providers-ftp==3.11.1
apache-airflow-providers-google==10.1.1
apache-airflow-providers-http==4.13.1
apache-airflow-providers-imap==3.7.0
apache-airflow-providers-smtp==1.8.0
apache-airflow-providers-sqlite==3.9.0
apispec==6.6.1
argcomplete==3.5.3
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
babel==2.17.0
backoff==2.2.1
billiard==4.2.1
black==25.1.0
blinker==1.8.2
cachelib==0.9.0
cachetools==5.5.2
cattrs==24.1.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clickclick==20.10.2
colorama==0.4.6
colorlog==6.9.0
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==41.0.7
db-dtypes==1.4.1
Deprecated==1.2.18
dill==0.3.1.1
dnspython==2.6.1
docstring_parser==0.16
email_validator==2.2.0
Flask==2.2.5
Flask-AppBuilder==4.5.0
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
flower==2.0.1
frozenlist==1.5.0
fsspec==2024.10.0
gcloud-aio-auth==4.2.3
gcloud-aio-bigquery==7.1.0
gcloud-aio-storage==9.2.0
google-ads==25.2.0
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.82.0
google-cloud-appengine-logging==1.6.0
google-cloud-audit-log==0.3.0
google-cloud-automl==2.16.2
google-cloud-bigquery==3.30.0
google-cloud-bigquery-datatransfer==3.19.0
google-cloud-bigtable==2.29.0
google-cloud-build==3.31.0
google-cloud-compute==1.26.0
google-cloud-container==2.56.0
google-cloud-core==2.4.2
google-cloud-datacatalog==3.25.1
google-cloud-dataflow-client==0.8.16
google-cloud-dataform==0.6.0
google-cloud-dataplex==2.7.1
google-cloud-dataproc==5.18.0
google-cloud-dataproc-metastore==1.18.1
google-cloud-dlp==3.28.0
google-cloud-kms==3.4.0
google-cloud-language==2.17.0
google-cloud-logging==3.11.4
google-cloud-memcache==1.12.0
google-cloud-monitoring==2.27.0
google-cloud-orchestration-airflow==1.17.0
google-cloud-os-login==2.17.0
google-cloud-pubsub==2.28.0
google-cloud-redis==2.18.0
google-cloud-resource-manager==1.14.1
google-cloud-secret-manager==2.23.1
google-cloud-spanner==3.52.0
google-cloud-speech==2.31.0
google-cloud-storage==2.19.0
google-cloud-tasks==2.19.1
google-cloud-texttospeech==2.25.0
google-cloud-translate==3.20.1
google-cloud-videointelligence==2.16.0
google-cloud-vision==3.10.0
google-cloud-workflows==1.17.0
google-crc32c==1.6.0
google-re2==1.1.20240702
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpc-google-iam-v1==0.14.1
grpc-interceptor==0.15.4
grpcio==1.65.5
grpcio-gcp==0.2.2
grpcio-status==1.62.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.0
humanize==4.10.0
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.4.5
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==0.10.0
json-merge-patch==0.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.2
lazy-object-proxy==1.10.0
limits==3.13.0
linkify-it-py==2.0.3
lockfile==0.12.2
looker-sdk==25.2.0
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.5.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.2.3
oauthlib==3.2.2
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
pandas-gbq==0.28.0
pathspec==0.12.1
pendulum==3.0.0
platformdirs==4.3.6
pluggy==1.5.0
prison==0.2.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.0
proto-plus==1.26.0
protobuf==4.25.6
psutil==6.1.1
pyarrow==19.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydata-google-auth==1.9.1
Pygments==2.19.1
PyJWT==2.9.0
pyOpenSSL==25.0.0
pyparsing==3.2.1
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.6.0
rpds-py==0.20.1
rsa==4.9
setproctitle==1.3.4
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.54
sqlalchemy-bigquery==1.12.1
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.4.0
text-unidecode==1.3
time-machine==2.15.0
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
uc-micro-py==1.0.3
universal_pathlib==0.2.6
uritemplate==4.1.1
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
Werkzeug==2.2.3
wirerope==1.0.0
wrapt==1.17.2
WTForms==3.1.2
yarl==1.15.2
zipp==3.20.2
